+++ include/copperfin/studio/document_model.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "copperfin/studio/designer_context.h"
 #include "copperfin/runtime/prg_engine.h"
 #include "copperfin/vfp/asset_inspector.h"
 #include "copperfin/vfp/dbf_table.h"
@@ -62,6 +63,7 @@ struct StudioDocumentModel {
     std::vector<runtime::PrgStaticDiagnostic> static_diagnostics;
     bool table_preview_available = false;
     vfp::DbfTable table_preview{};
+    std::vector<StudioDesignerContextResult> designer_contexts{};
 };
 
 struct StudioPropertySnapshot {